+++ .github/workflows/check_style.yml
@@ -1,41 +1,41 @@
-name: Java CI with Maven
-
-on:
-  pull_request:
-    branches: 
-      - main
-
-jobs:
-  build:
-
-    runs-on: ubuntu-22.04
-
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up JDK 11
-      uses: actions/setup-java@v4
-      with:
-        java-version: '11'
-        distribution: 'temurin'
-        cache: maven
-
-    - name: Clean Maven
-      run: mvn clean
-
-    - name: Check style java
-      id: check-style
-      run: mvn checkstyle:check
-      continue-on-error: true
-
-    - name: Generate check style report
-      if: ${{ steps.check-style.outcome == 'failure' }}
-      run: mvn site
-
-    - name: Upload check style result
-      if: ${{ steps.check-style.outcome == 'failure' }}
-      uses: actions/upload-artifact@v4
-      with:
-        name: Check style result
-        path: ./target/site/**
-        retention-days: 3
-
+#name: Java CI with Maven
+#
+#on:
+#  pull_request:
+#    branches:
+#      - main
+#
+#jobs:
+#  build:
+#
+#    runs-on: ubuntu-22.04
+#
+#    steps:
+#    - uses: actions/checkout@v4
+#    - name: Set up JDK 11
+#      uses: actions/setup-java@v4
+#      with:
+#        java-version: '11'
+#        distribution: 'temurin'
+#        cache: maven
+#
+#    - name: Clean Maven
+#      run: mvn clean
+#
+#    - name: Check style java
+#      id: check-style
+#      run: mvn checkstyle:check
+#      continue-on-error: true
+#
+#    - name: Generate check style report
+#      if: ${{ steps.check-style.outcome == 'failure' }}
+#      run: mvn site
+#
+#    - name: Upload check style result
+#      if: ${{ steps.check-style.outcome == 'failure' }}
+#      uses: actions/upload-artifact@v4
+#      with:
+#        name: Check style result
+#        path: ./target/site/**
+#        retention-days: 3
+#
