+++ src/main/java/com/agest/model/sele2/GlobalSetting.java
@@ -1,4 +1,4 @@
-package com.agest.model;
+package com.agest.model.sele2;
 
 import lombok.Getter;
