+++ src/main/java/com/agest/model/sele2/Series.java
@@ -1,4 +1,4 @@
-package com.agest.model;
+package com.agest.model.sele2;
 
 import lombok.Getter;
 import org.apache.commons.lang3.RandomUtils;
