+++ src/main/java/com/agest/page/CreatePanelPage.java
@@ -1,7 +1,7 @@
 package com.agest.page;
 
-import com.agest.model.Panel;
-import com.agest.model.Series;
+import com.agest.model.sele2.Panel;
+import com.agest.model.sele2.Series;
 import com.codeborne.selenide.SelenideElement;
 import io.qameta.allure.Step;
