+++ src/main/java/com/agest/page/PanelManagerPage.java
@@ -1,6 +1,6 @@
 package com.agest.page;
 
-import com.agest.model.Panel;
+import com.agest.model.sele2.Panel;
 import com.agest.utils.AlertUtils;
 import com.codeborne.selenide.SelenideElement;
 import io.qameta.allure.Step;
