+++ src/test/java/sele2/DataProfile067.java
@@ -1,26 +1,29 @@
-import com.agest.model.User;
+package sele2;
+
+import com.agest.model.sele2.User;
 import com.agest.page.DashBoardPage;
 import com.agest.page.DataProfilePage;
 import com.agest.page.LoginPage;
 import com.agest.utils.Constants;
-import com.agest.utils.JsonUtils;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
+import java.util.Collections;
 import java.util.List;
 
-public class DataProfile065 extends TestBase {
+public class DataProfile067 extends TestBase {
     private final User user = User.getUser2();
     private final LoginPage loginPage = new LoginPage();
     private final DashBoardPage dashBoardPage = new DashBoardPage();
     private final DataProfilePage dataProfilePage = new DataProfilePage();
-    private final List<String> presetDataProfiles = JsonUtils.toList(Constants.PRESET_DATA_PROFILES, String.class);
 
-    @Test(description = "Verify that all Pre-set Data Profiles are populated correctly")
-    public void verifyPresetDataProfilesArePopulatedCorrectly() {
+    @Test(description = "Verify that Data Profiles are listed alphabetically")
+    public void verifyDataProfilesAreListedAlphabetically() {
         loginPage.login(user, Constants.TEST_REPO);
         dashBoardPage.openDataProfiles();
         List<String> actualDataProfiles = dataProfilePage.getDataProfiles();
-        Assert.assertEquals(presetDataProfiles, actualDataProfiles, "The data not match with preset!");
+        List<String> expectedDataProfiles = dataProfilePage.getDataProfiles();
+        Collections.sort(expectedDataProfiles);
+        Assert.assertEquals(actualDataProfiles, expectedDataProfiles, "The data is not listed alphabetically!");
     }
 }
