+++ src/test/java/sele2/Login001.java
@@ -1,4 +1,6 @@
-import com.agest.model.User;
+package sele2;
+
+import com.agest.model.sele2.User;
 import com.agest.page.DashBoardPage;
 import com.agest.page.LoginPage;
 import org.testng.annotations.Test;
